+++ packages/core/styles/src/types/index.ts
@@ -32,3 +32,5 @@ export type AlignItems =
   | "baseline";
 export type TextAlign = "left" | "right" | "center" | "justify";
 export type TextDecoration = "none" | "underline";
+export type Position = "absolute" | "fixed" | "relative" | "static" | "sticky";
+export type Overflow = "visible" | "hidden" | "scroll" | "auto";