+++ packages/react/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This package is intended for internal use in generating builds of each design system package. It contains all the necessary settings and dependencies to optimize the creation of our builds.
 
+## 2023-01-27 `2.2.0`
+
+### 🎉 New features
+
+- Added new options to `appearance` property on `Popover` component. ([#76](https://github.com/TiendaNube/nimbus-design-system/pull/76) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new options to `appearance` property on `Link` component. ([#76](https://github.com/TiendaNube/nimbus-design-system/pull/76) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new properties to `Box` component. ([#76](https://github.com/TiendaNube/nimbus-design-system/pull/76) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-01-09 `2.1.0`
 
 ### 🎉 New features